Reduce dependency to Xiliary components
Some bundles use a few classes of the Xiliary [1] utilities.
The maintenance effort to include these dependencies into the
p2 repository does not justify the benefit of sharing such a
small amount of code.

Replace the uses of Xiliary with tailored copies and remove the
build steps that include Xiliary bundles.

[1] https://github.com/fappel/xiliary